Remove boundary submission logic
