+++ crates/driver/src/boundary/mod.rs
@@ -23,7 +23,6 @@
 //! Software (2014)
 
 pub mod liquidity;
-pub mod mempool;
 pub mod quote;
 pub mod score;
 pub mod settlement;
@@ -34,7 +33,6 @@ use {crate::infra::blockchain::Ethereum, url::Url};
 pub use {
     anyhow::{Error, Result},
     contracts,
-    mempool::Mempool,
     model::order::OrderData,
     settlement::Settlement,
     shared::ethrpc::Web3,
