+++ crates/driver/src/tests/setup/solver.rs
@@ -209,7 +209,6 @@ impl Solver {
                     max_confirm_time: Default::default(),
                     retry_interval: Default::default(),
                     kind: infra::mempool::Kind::Public(infra::mempool::RevertProtection::Disabled),
-                    submission: infra::mempool::SubmissionLogic::Native,
                 }],
             )
             .await
